clean up painter plugin
