+++ substance/SubstancePainter/Scripts/Prism_SubstancePainter_Variables.py
@@ -9,12 +9,12 @@ class Prism_SubstancePainter_Variables(object):
         self.appShortName = "SubstancePainter"
         self.appType = "3d"
         self.hasQtParent = True
-        self.sceneFormats = [".format"]
+        self.sceneFormats = [".spp", ".exr", ".png", ".jpg", "tiff"]
         self.appSpecificFormats = self.sceneFormats
-        self.outputFormats = [".abc", ".obj", ".fbx", "ShotCam"]
+        self.outputFormats = [".exr", ".png", ".jpg", "tiff"]
         self.appColor = [255, 255, 255]
         self.renderPasses = []
+        self.hasFrameRange = False
         self.platforms = ["Windows"]
-        self.pluginDirectory = os.path.abspath(
-            os.path.dirname(os.path.dirname(__file__))
-        )
+        self.pluginDirectory = os.path.abspath(os.path.dirname(os.path.dirname(__file__)))
+        self.appIcon = os.path.join(self.pluginDirectory, "Resources", "Substance.ico")
