--- substance/UI/ExportTextureUI/exporttexturemainwindow.cpp
@@ -1,14 +1,0 @@
-#include "exporttexturemainwindow.h"
-#include "./ui_exporttexturemainwindow.h"
-
-ExportTextureMainWindow::ExportTextureMainWindow(QWidget *parent)
-    : QMainWindow(parent)
-    , ui(new Ui::ExportTextureMainWindow)
-{
-    ui->setupUi(this);
-}
-
-ExportTextureMainWindow::~ExportTextureMainWindow()
-{
-    delete ui;
-}
