--- substance/UI/ExportTextureUI/main.cpp
@@ -1,23 +1,0 @@
-#include "exporttexturemainwindow.h"
-
-#include <QApplication>
-#include <QLocale>
-#include <QTranslator>
-
-int main(int argc, char *argv[])
-{
-    QApplication a(argc, argv);
-
-    QTranslator translator;
-    const QStringList uiLanguages = QLocale::system().uiLanguages();
-    for (const QString &locale : uiLanguages) {
-        const QString baseName = "ExportTextureUI_" + QLocale(locale).name();
-        if (translator.load(":/i18n/" + baseName)) {
-            a.installTranslator(&translator);
-            break;
-        }
-    }
-    ExportTextureMainWindow w;
-    w.show();
-    return a.exec();
-}
